requirements.txt
fastapi
uvicorn
pillow
python-multipart
python-jose[cryptography]
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
sqlalchemy>=2.0
deepface
tf-keras
numpy
scikit-learn
